+++ gcip2/pipeline_vars/__init__.py
@@ -34,3 +34,5 @@ class Predefined(VarEnum, enum.Enum):
     CI_COMMIT_BRANCH = enum.auto()
     CI_DEFAULT_BRANCH = enum.auto()
     CI_PIPELINE_SOURCE = enum.auto()
+    CI_JOB_IMAGE = enum.auto()
+    CI_COMMIT_TAG = enum.auto()