+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 ## Changelog
 
+### [30.07.26 = 0.0.5]
+- added base project structure initialization with `gcip2 init`
+- added base predefined CI variables for pipelines
+- added pytest and integration regression tests with golden pipelines
+- removed apply method from JobBuilderImpl on build
+- removed duplicate of PipelineBuilderImpl
+- updated documentation
+- fixed some bugs and errors
+
 ### [28.07.26 - 0.0.4]
 - added GitlabCiBuilderImpl
 - added BaseLinux and _base param for JobBuilderImpl